% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_raster_stack_fun.R
\name{env_raster_stack}
\alias{env_raster_stack}
\title{Creates raster stacks from a environmental covariate layer and home range}
\usage{
env_raster_stack(env, hrr)
}
\arguments{
\item{env}{A environmental raster layer}

\item{hrr}{A home range that the environemntal covariate is resampled to.}
}
\value{
The ouput is a raster stack that contains the home ranges and the
environmental covariate layer.
}
\description{
Creates raster stacks from a environmental covariate layer and home range
}
