% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iso2Raster_fun.R
\name{iso2raster}
\alias{iso2raster}
\title{Creates rasters from the LoCoH isopleths}
\usage{
iso2raster(isopleth, resolution)
}
\arguments{
\item{isopleth}{A isopleth created from LoCoH home range estimates.}

\item{resolution}{Specify a resolution of the resulting raster.}
}
\value{
The output is a list of rasters for individual each within a given
interval time frame.
}
\description{
Creates rasters from the LoCoH isopleths
}
